fix: 대쉬보드 컨트롤러 로직 수정 S13P31A305-341

+++ backend/ongil/src/main/java/kr/co/ongil/domain/patient/dashboard/dto/CallStatisticsDto.java
@@ -10,6 +10,6 @@ import lombok.Setter;
 @NoArgsConstructor
 @AllArgsConstructor
 public class CallStatisticsDto {
-    private Integer patientId;
-    private Integer callCount;
+    private Long patientId;
+    private Long callCount;
 }
